requirements.txt
librosa
noisereduce
soundfile
numpy
matplotlib.pyplot
os
scipy.io
moviepy.editor